Extract workspace flow types into shared module
