+++ src/backend/domains/workspace/state/flow-state.ts
@@ -1,14 +1,9 @@
 import { CIStatus, PRState, RatchetState } from '@/shared/core';
+import type { WorkspaceCiObservation, WorkspaceFlowPhase } from '@/shared/workspace-flow-state';
 
 const CI_UNKNOWN_GRACE_MS = 90_000;
 
-export type WorkspaceFlowPhase =
-  | 'NO_PR'
-  | 'CI_WAIT'
-  | 'RATCHET_VERIFY'
-  | 'RATCHET_FIXING'
-  | 'READY'
-  | 'MERGED';
+export type { WorkspaceCiObservation, WorkspaceFlowPhase } from '@/shared/workspace-flow-state';
 
 export interface WorkspaceFlowStateInput {
   prUrl: string | null;
@@ -23,14 +18,6 @@ export type WorkspaceFlowStateSource = Pick<
   WorkspaceFlowStateInput,
   'prUrl' | 'prState' | 'prCiStatus' | 'prUpdatedAt' | 'ratchetEnabled' | 'ratchetState'
 >;
-
-export type WorkspaceCiObservation =
-  | 'NOT_FETCHED'
-  | 'NO_CHECKS'
-  | 'CHECKS_PENDING'
-  | 'CHECKS_FAILED'
-  | 'CHECKS_PASSED'
-  | 'CHECKS_UNKNOWN';
 
 export interface WorkspaceFlowState {
   phase: WorkspaceFlowPhase;
